Apply renaming for suggestion
