+++ src/app.tsx
@@ -3,14 +3,14 @@ import {
   RouterProvider,
   createBrowserRouter,
 } from "react-router-dom";
-import "./App.css";
-import Layout from "./Pages/Layout";
-import { PATHNAMES } from "./Services/RouteService";
-import { ProtectedRoute } from "./Routes/ProtectedRoute";
-import Home from "./Pages/Home";
-import Waiter from "./Pages/Waiter";
-import Admin from "./Pages/Admin";
-import Login from "./Pages/Login";
+import "./app.css";
+import Layout from "./pages/layout";
+import { PATHNAMES } from "./services/route-service";
+import { ProtectedRoute } from "./routes/protected-route";
+import Home from "./pages/home";
+import Waiter from "./pages/waiter";
+import Admin from "./pages/admin";
+import Login from "./pages/login";
 
 const ROUTES: RouteObject[] = [
   {
