+++ src/routes/protected-route.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from "react";
 import { Navigate, Outlet, useNavigate } from "react-router-dom";
-import { getSession } from "../Services/TokenRepository";
-import { getUser } from "../Services/UserRepository";
-import { PATHNAMES } from "../Services/RouteService";
+import { getSession } from "../services/token-repository";
+import { getUser } from "../services/user-repository";
+import { PATHNAMES } from "../services/route-service";
 
 export const ProtectedRoute = () => {
   const { token, userId } = getSession();
